worked on script.js
the website will generate buttons from local storage when the page is reloaded. #
There is an error where data is overridden, due to cityNumber being used to populate localstorage cityNumber items.
A work around could be an if statement to use either cityNumber the array or 'cities' from local storage.
Or easier way would be to just populate one array in local storage that keeps track of all entries.

TODO
Still to work on getting query results on the page.

SUCCESS
Slight improvements are being made, and better understanding is gained through continued working it out.